Update README to limit project options to a single CSS module.
Looks like, as is, supporting multiple CSS modules w/ the Closure Stylesheets compiler would be tricky (because of renaming, it would be hard to compile separately). May try to hack around this limitation in later versions, but it would be nicer if the GSS compiler supported modules the same way the JS compiler does.